Base code complete

+++ pellets.py
@@ -8,13 +8,13 @@ class Pellet(object):
     def __init__(self, row, column):
         """Initialize class variables"""
         self.name = PELLET
-        self.position = Vector2(column * TILEWIDTH, row * TILEHEIGHT)
+        self.position = Vector2(column*TILEWIDTH, row*TILEHEIGHT)
         self.color = WHITE
         self.radius = int(2 * TILEWIDTH / 16)
-        self.collideRadius = int(2 * TILEWIDTH / 16)
+        self.collideRadius = 2 * TILEWIDTH / 16
         self.points = 10
         self.visible = True
-
+        
     def render(self, screen):
         """Draws Pellets onto the screen"""
         if self.visible:
@@ -31,8 +31,8 @@ class PowerPellet(Pellet):
         self.radius = int(8 * TILEWIDTH / 16)
         self.points = 50
         self.flashTime = 0.2
-        self.timer = 0
-
+        self.timer= 0
+        
     def update(self, dt):
         """Game loop called once per frame of the game"""
         self.timer += dt
@@ -54,7 +54,7 @@ class PelletGroup(object):
         """Game loop called once per frame of the game"""
         for powerpellet in self.powerpellets:
             powerpellet.update(dt)
-
+                
     def createPelletList(self, pelletfile):
         """Creates a list of Pellets"""
         data = self.readPelletfile(pelletfile)
@@ -66,17 +66,17 @@ class PelletGroup(object):
                     pp = PowerPellet(row, col)
                     self.pelletList.append(pp)
                     self.powerpellets.append(pp)
-
+                    
     def readPelletfile(self, textfile):
         """Reads in text file"""
         return np.loadtxt(textfile, dtype='<U1')
-
+    
     def isEmpty(self):
         """Checks to see when the pelletList is empty"""
         if len(self.pelletList) == 0:
             return True
         return False
-
+    
     def render(self, screen):
         """Draws each Pellet onto the screen"""
         for pellet in self.pelletList:
